@@ -7,6 +7,7 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
+    "build:raiden-package": "vue-cli-service build --mode raiden-package",
     "lint": "vue-cli-service lint",
     "i18n:report": "vue-cli-service i18n:report --src './src/**/*.?(js|vue)' --locales './src/locales/**/*.json'",
     "lint:check": "vue-cli-service lint --no-fix --max-warnings 0",
@@ -14,57 +15,58 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
-    "eslint-config-prettier": "^6.0.0",
+    "core-js": "^2.6.9",
     "ethereum-blockies-base64": "^1.0.2",
     "ethers": "^4.0.37",
-    "lodash": "^4.17.14",
+    "lodash": "^4.17.15",
     "raiden-ts": "file:../raiden-ts",
     "register-service-worker": "^1.6.2",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.3",
+    "tiny-async-pool": "^1.0.4",
     "vue": "^2.6.10",
     "vue-class-component": "^7.1.0",
-    "vue-i18n": "^8.12.0",
-    "vue-property-decorator": "^8.2.1",
-    "vue-router": "^3.0.7",
-    "vuetify": "2.0.4",
+    "vue-i18n": "^8.14.1",
+    "vue-property-decorator": "^8.2.2",
+    "vue-router": "^3.1.3",
+    "vue-virtual-scroller": "^1.0.0-rc.2",
+    "vuetify": "2.0.19",
     "vuex": "^3.1.1"
   },
   "devDependencies": {
     "@kazupon/vue-i18n-loader": "^0.4.1",
-    "@types/jest": "^24.0.15",
-    "@types/lodash": "^4.14.136",
+    "@types/jest": "^24.0.18",
+    "@types/lodash": "^4.14.141",
+    "@types/tiny-async-pool": "^1.0.0",
     "@types/vue-i18n": "^7.0.0",
-    "@types/webpack": "^4.4.0",
-    "@vue/cli": "^3.10.0",
-    "@vue/cli-plugin-babel": "^3.10.0",
-    "@vue/cli-plugin-eslint": "^3.10.0",
-    "@vue/cli-plugin-pwa": "^3.10.0",
-    "@vue/cli-plugin-typescript": "^3.10.0",
-    "@vue/cli-plugin-unit-jest": "^3.10.0",
-    "@vue/cli-service": "^3.10.0",
+    "@types/webpack": "^4.39.2",
+    "@vue/cli": "^3.11.0",
+    "@vue/cli-plugin-babel": "^3.11.0",
+    "@vue/cli-plugin-eslint": "^3.11.0",
+    "@vue/cli-plugin-pwa": "^3.11.0",
+    "@vue/cli-plugin-typescript": "^3.11.0",
+    "@vue/cli-plugin-unit-jest": "^3.11.0",
+    "@vue/cli-service": "^3.11.0",
     "@vue/eslint-config-prettier": "^5.0.0",
     "@vue/eslint-config-typescript": "^4.0.0",
     "@vue/test-utils": "^1.0.0-beta.29",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.2",
+    "babel-eslint": "^10.0.3",
     "eslint": "^5.16.0",
-    "eslint-plugin-jsdoc": "^15.8.3",
-    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-jsdoc": "^15.9.5",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-vue": "^5.2.3",
     "eslint-plugin-vue-i18n": "^0.2.3",
-    "eslint-plugin-vuetify": "^1.0.0-beta.3",
+    "eslint-plugin-vuetify": "^1.0.0-beta.4",
     "flush-promises": "^1.0.2",
-    "jest-junit": "^6.4.0",
-    "sass": "^1.22.9",
-    "sass-loader": "^7.1.0",
+    "jest-junit": "^8.0.0",
+    "sass": "^1.22.12",
+    "sass-loader": "^7.3.1",
     "source-map-loader": "^0.2.4",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^3.0.2",
     "ts-jest": "^23.10.5",
-    "typescript": "~3.3.4000",
+    "typescript": "~3.5.3",
     "vue-cli-plugin-i18n": "^0.6.0",
-    "vue-cli-plugin-vuetify": "^0.6.1",
-    "vue-jest": "^3.0.4",
+    "vue-cli-plugin-vuetify": "^0.6.3",
+    "vue-jest": "^3.0.5",
     "vue-template-compiler": "^2.6.10",
     "vuetify-loader": "^1.3.0"
   },
