@@ -1,6 +1,7 @@
 import { Observable } from 'rxjs';
-import { filter, scan, startWith } from 'rxjs/operators';
+import { filter, scan, startWith, share } from 'rxjs/operators';
 import { isActionOf } from 'typesafe-actions';
+import { memoize } from 'lodash';
 
 import { RaidenAction } from '../actions';
 import { Presences } from './types';
@@ -9,24 +10,24 @@ import { matrixPresenceUpdate } from './actions';
 /**
  * Helper to map/get an aggregated Presences observable from action$ bus
  * Known presences as { address: <last seen MatrixPresenceUpdateAction> } mapping
- * As this helper is basically a scan/reduce, you can't simply startWith the first/initial value,
- * as it needs to also be the initial mapping for the scan itself, so instead of pipe+startWith,
- * as we usually do with state$, we need to get the initial value as parameter when it's used in
- * withLatestFrom in some inner observable
+ * It's memoized and shared, so all subscriptions share the same mapped/output object
  *
  * @param action$ - Observable
  * @returns Observable of aggregated Presences from subscription to now
  */
-export const getPresences$ = (action$: Observable<RaidenAction>): Observable<Presences> =>
-  action$.pipe(
-    filter(isActionOf(matrixPresenceUpdate)),
-    scan(
-      // scan all presence update actions and populate/output a per-address mapping
-      (presences, update) => ({
-        ...presences,
-        [update.meta.address]: update,
-      }),
-      {},
+export const getPresences$ = memoize(
+  (action$: Observable<RaidenAction>): Observable<Presences> =>
+    action$.pipe(
+      filter(isActionOf(matrixPresenceUpdate)),
+      scan(
+        // scan all presence update actions and populate/output a per-address mapping
+        (presences, update) => ({
+          ...presences,
+          [update.meta.address]: update,
+        }),
+        {} as Presences,
+      ),
+      share(),
+      startWith({}),
     ),
-    startWith({}),
-  );
+);
