@@ -10,35 +10,6 @@ import { Tokens } from './types';
 
 Vue.use(Router);
 
-/**
- * Checks whether a token parameter is present and whether
- * requested token network address exists in store.
- */
-export const checkTokenNetworkRoute = (
-  { token }: Dictionary<string>,
-  next: (to?: string | void) => void,
-  tokens: Tokens
-): void => {
-  if (!token) return next();
-
-  return Object.keys(tokens)
-    .map(key => tokens[key].address === token)
-    .includes(true)
-    ? next()
-    : next('/');
-};
-
-/**
- * Decorator that wraps vue router's `NavigationGuard` and
- * the current list of tokens in the store. This makes testing
- * `checkTokenNetworkRoute` very straight forward.
- */
-export const tokenNetworkGuard = (
-  { params }: Route,
-  _: Route,
-  next: (to?: string | void) => void
-) => checkTokenNetworkRoute(params, next, store.getters.allTokens);
-
 const router = new Router({
   mode: 'history',
   base: process.env.BASE_URL,
@@ -61,8 +32,7 @@ const router = new Router({
       meta: {
         title: 'Send'
       },
-      component: () => import('./views/Payment.vue'),
-      beforeEnter: tokenNetworkGuard
+      component: () => import('./views/Payment.vue')
     },
     {
       path: '/connect',
@@ -78,8 +48,7 @@ const router = new Router({
       meta: {
         title: 'Select hub'
       },
-      component: () => import('./views/SelectHub.vue'),
-      beforeEnter: tokenNetworkGuard
+      component: () => import('./views/SelectHub.vue')
     },
     {
       path: '/connect/:token/:partner',
@@ -87,8 +56,7 @@ const router = new Router({
       meta: {
         title: 'Open channel'
       },
-      component: () => import('./views/OpenChannel.vue'),
-      beforeEnter: tokenNetworkGuard
+      component: () => import('./views/OpenChannel.vue')
     },
     {
       path: '/channels/:token',
@@ -96,8 +64,7 @@ const router = new Router({
       meta: {
         title: 'Channels'
       },
-      component: () => import('./views/Channels.vue'),
-      beforeEnter: tokenNetworkGuard
+      component: () => import('./views/Channels.vue')
     }
   ]
 });
