@@ -2,6 +2,7 @@ jest.mock('@/services/raiden-service');
 jest.useFakeTimers();
 
 import flushPromises from 'flush-promises';
+import { $identicon } from '../utils/mocks';
 import store from '@/store';
 import { mount, Wrapper } from '@vue/test-utils';
 import AddressInput from '@/components/AddressInput.vue';
@@ -16,30 +17,32 @@ Vue.use(Vuetify);
 describe('AddressInput', function() {
   let wrapper: Wrapper<AddressInput>;
   let raiden: Mocked<RaidenService>;
-  let mockIdenticon: jest.Mock<any, any>;
+
   const excludeAddress: string = '0x65E84e07dD79F3f03d72bc0fab664F56E6C55909';
   const blockAddress: string = '0x123456789009876543211234567890';
 
-  beforeEach(() => {
-    mockIdenticon = jest.fn().mockResolvedValue('');
-    raiden = new RaidenService(store) as Mocked<RaidenService>;
-    wrapper = mount(AddressInput, {
+  function vueFactory(value: string = '', excluded?: string, blocked?: string) {
+    return mount(AddressInput, {
+      sync: false,
       propsData: {
-        value: '',
-        exclude: [excludeAddress],
-        block: [blockAddress]
+        value,
+        exclude: excluded ? [excluded] : undefined,
+        block: blocked ? [blocked] : undefined
       },
       mocks: {
         $raiden: raiden,
-        $identicon: {
-          getIdenticon: mockIdenticon
-        },
+        $identicon: $identicon(),
         $t: (msg: string) => msg
       }
     });
+  }
+
+  beforeEach(() => {
+    raiden = new RaidenService(store) as Mocked<RaidenService>;
   });
 
   it('should show no validation messages', () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     const messages = wrapper.find('.v-messages__message');
     expect(wrapper.props().value).toBe('');
     expect(messages.exists()).toBe(true);
@@ -47,6 +50,7 @@ describe('AddressInput', function() {
   });
 
   it('should show a this address cannot be an empty message', async () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     mockInput(wrapper, '0x21b');
     await wrapper.vm.$nextTick();
     mockInput(wrapper);
@@ -62,6 +66,7 @@ describe('AddressInput', function() {
   });
 
   it('should should show a no valid address message', async () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     mockInput(wrapper, '0x21b');
     await wrapper.vm.$nextTick();
 
@@ -71,6 +76,7 @@ describe('AddressInput', function() {
   });
 
   it('should should show a not checksum format message if address not in checksum format', async () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     mockInput(wrapper, '0x774afb0652ca2c711fd13e6e9d51620568f6ca82');
     await wrapper.vm.$nextTick();
 
@@ -80,6 +86,7 @@ describe('AddressInput', function() {
   });
 
   test('valid checksum address should fire input event', async () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     mockInput(wrapper, '0x1D36124C90f53d491b6832F1c073F43E2550E35b');
     await wrapper.vm.$nextTick();
 
@@ -90,12 +97,15 @@ describe('AddressInput', function() {
   });
 
   test('setting a valid address should render a blockie', async () => {
+    wrapper = vueFactory('', excludeAddress, blockAddress);
     wrapper.setProps({ value: '0x1D36124C90f53d491b6832F1c073F43E2550E35b' });
-    expect(mockIdenticon).toHaveBeenCalled();
+    await wrapper.vm.$nextTick();
+    expect(wrapper.vm.$identicon.getIdenticon).toHaveBeenCalled();
   });
 
   describe('resolving ens names', () => {
     test('successfully resolved', async () => {
+      wrapper = vueFactory('', excludeAddress, blockAddress);
       raiden.ensResolve = jest
         .fn()
         .mockResolvedValue('0x1D36124C90f53d491b6832F1c073F43E2550E35b');
@@ -120,6 +130,7 @@ describe('AddressInput', function() {
     });
 
     test('could not resolve an address', async () => {
+      wrapper = vueFactory('', excludeAddress, blockAddress);
       raiden.ensResolve = jest.fn().mockResolvedValue(null);
 
       mockInput(wrapper, 'enstest.test');
@@ -138,6 +149,7 @@ describe('AddressInput', function() {
     });
 
     test('failed to resolve an address', async () => {
+      wrapper = vueFactory('', excludeAddress, blockAddress);
       raiden.ensResolve = jest
         .fn()
         .mockRejectedValue(Error('something went wrong'));
@@ -160,6 +172,7 @@ describe('AddressInput', function() {
 
   describe('exclude & block address', () => {
     it('should show error message if excluded address is entered', async () => {
+      wrapper = vueFactory('', excludeAddress, blockAddress);
       mockInput(wrapper, excludeAddress);
       await wrapper.vm.$nextTick();
 
@@ -171,6 +184,7 @@ describe('AddressInput', function() {
     });
 
     it('should show error message if blocked address is entered', async () => {
+      wrapper = vueFactory('', excludeAddress, blockAddress);
       mockInput(wrapper, blockAddress);
       await wrapper.vm.$nextTick();
 
@@ -180,24 +194,45 @@ describe('AddressInput', function() {
     });
 
     it('should not show error message if there is no exclude or block prop', async () => {
-      wrapper = mount(AddressInput, {
-        propsData: {
-          value: ''
-        },
-        mocks: {
-          $raiden: raiden,
-          $identicon: {
-            getIdenticon: mockIdenticon
-          },
-          $t: (msg: string) => msg
-        }
-      });
+      wrapper = vueFactory();
 
       mockInput(wrapper, excludeAddress);
       await wrapper.vm.$nextTick();
 
       const messages = wrapper.find('.v-messages__message');
       expect(messages.exists()).toBe(false);
+    });
+  });
+
+  describe('internal value should react to model changes', () => {
+    test('invalid value should not update address', () => {
+      wrapper = vueFactory('0xsdajlskdj');
+      expect(wrapper.vm.$data.address).toBe('');
+    });
+
+    test('valid value should update address', () => {
+      wrapper = vueFactory('0x1D36124C90f53d491b6832F1c073F43E2550E35b');
+      expect(wrapper.vm.$data.address).toBe(
+        '0x1D36124C90f53d491b6832F1c073F43E2550E35b'
+      );
+    });
+
+    test('address should be updated on valid changes', async () => {
+      wrapper = vueFactory();
+      expect(wrapper.vm.$data.address).toBe('');
+      wrapper.setProps({ value: '0x1D36124C90f53d491b6832F1c073F43E2550E35b' });
+      await wrapper.vm.$nextTick();
+      expect(wrapper.vm.$data.address).toBe(
+        '0x1D36124C90f53d491b6832F1c073F43E2550E35b'
+      );
+    });
+
+    test('address should not be updated on invalid changes', async () => {
+      wrapper = vueFactory();
+      expect(wrapper.vm.$data.address).toBe('');
+      wrapper.setProps({ value: '0x1aaaaaadshjd' });
+      await wrapper.vm.$nextTick();
+      expect(wrapper.vm.$data.address).toBe('');
     });
   });
 });
