@@ -1,6 +1,6 @@
 import { createLocalVue, mount, shallowMount, Wrapper } from '@vue/test-utils';
 import { addElemWithDataAppToBody } from '../utils/dialog';
-import Vuex, { Store } from 'vuex';
+import Vuex from 'vuex';
 import Vuetify from 'vuetify';
 import { TestData } from '../data/mock-data';
 import SelectHub from '@/views/SelectHub.vue';
@@ -9,13 +9,17 @@ import { mockInput } from '../utils/interaction-utils';
 import Mocked = jest.Mocked;
 import { RouteNames } from '@/route-names';
 import Vue from 'vue';
+import store from '@/store';
+import { $identicon } from '../utils/mocks';
+import flushPromises from 'flush-promises';
 
 Vue.use(Vuetify);
 
 describe('SelectHub.vue', function() {
+  addElemWithDataAppToBody();
+
   let wrapper: Wrapper<SelectHub>;
   let router: Mocked<VueRouter>;
-  let callArgs: () => any;
   let vuetify: typeof Vuetify;
 
   const testToken = (address: string) =>
@@ -23,30 +27,19 @@ describe('SelectHub.vue', function() {
       address: address
     });
 
-  function createWrapper(
-    route: Route,
-    getter: any,
-    token: any,
-    shallow: boolean = false
-  ) {
+  function createWrapper(route: Route, token: any, shallow: boolean = false) {
     const localVue = createLocalVue();
     localVue.use(Vuex);
     const options = {
       vuetify,
       localVue,
-      store: new Store({
-        getters: {
-          token: jest.fn().mockReturnValue(() => getter)
-        }
-      }),
+      store,
       mocks: {
         $route: route,
         $router: router,
-        $identicon: {
-          getIdenticon: jest.fn().mockResolvedValue('')
-        },
+        $identicon: $identicon(),
         $raiden: {
-          getToken: jest.fn().mockResolvedValue(token)
+          fetchTokenData: jest.fn().mockResolvedValue(null)
         },
         $t: (msg: string) => msg
       }
@@ -58,43 +51,44 @@ describe('SelectHub.vue', function() {
   }
 
   beforeEach(() => {
-    addElemWithDataAppToBody();
     router = new VueRouter() as Mocked<VueRouter>;
     router.push = jest.fn().mockResolvedValue(null);
-    callArgs = () => router.push.mock.calls[0][0];
+  });
+
+  beforeEach(() => {
+    store.commit('reset');
   });
 
   test('when select hub is clicked with ', async () => {
+    const tokenAddress = '0xc778417E063141139Fce010982780140Aa0cD5Ab';
     const route = TestData.mockRoute({
-      token: '0xc778417E063141139Fce010982780140Aa0cD5Ab'
+      token: tokenAddress
     });
-    const token = testToken('0xc778417E063141139Fce010982780140Aa0cD5Ab');
-    wrapper = createWrapper(route, token, token);
+    const token = testToken(tokenAddress);
+    store.commit('updateTokens', { [tokenAddress]: token });
+    wrapper = createWrapper(route, token);
     mockInput(wrapper, '0x1D36124C90f53d491b6832F1c073F43E2550E35b');
     await wrapper.vm.$nextTick();
+    await flushPromises();
     wrapper.find('.action-button__button').trigger('click');
     expect(router.push).toHaveBeenCalledTimes(1);
-    expect(callArgs().name).toEqual(RouteNames.OPEN_CHANNEL);
+    expect(router.push).toHaveBeenCalledWith(
+      expect.objectContaining({ name: RouteNames.OPEN_CHANNEL })
+    );
   });
 
   test('when token address is not checksum should navigate to home', async () => {
     const route = TestData.mockRoute({
       token: '0xtoken'
     });
-    wrapper = createWrapper(route, undefined, undefined, true);
-    expect(router.push).toHaveBeenCalledTimes(1);
-    expect(callArgs().name).toEqual(RouteNames.HOME);
-  });
-
-  test('when token is not cached it should be retrieved', async () => {
-    const route = TestData.mockRoute({
-      token: '0xc778417E063141139Fce010982780140Aa0cD5Ab'
-    });
-    const token = testToken('0xc778417E063141139Fce010982780140Aa0cD5Ab');
-    wrapper = createWrapper(route, null, token, true);
+    wrapper = createWrapper(route, undefined, true);
     await wrapper.vm.$nextTick();
-    expect(wrapper.vm.$data.token.address).toEqual(
-      '0xc778417E063141139Fce010982780140Aa0cD5Ab'
+    await flushPromises();
+    expect(router.push).toHaveBeenCalledTimes(1);
+    expect(router.push).toHaveBeenCalledWith(
+      expect.objectContaining({
+        name: RouteNames.HOME
+      })
     );
   });
 
@@ -103,9 +97,16 @@ describe('SelectHub.vue', function() {
       token: '0xc778417E063141139Fce010982780140Aa0cD5Ab'
     });
 
-    wrapper = createWrapper(route, null, null, true);
+    wrapper = createWrapper(route, null, true);
+
     await wrapper.vm.$nextTick();
+    await flushPromises();
+
     expect(router.push).toHaveBeenCalledTimes(1);
-    expect(callArgs().name).toEqual(RouteNames.HOME);
+    expect(router.push).toHaveBeenCalledWith(
+      expect.objectContaining({
+        name: RouteNames.HOME
+      })
+    );
   });
 });
