@@ -1,7 +1,6 @@
-import { BigNumber } from 'ethers/utils';
 import { createStandardAction } from 'typesafe-actions';
 
-import { Address, Hash } from '../utils/types';
+import { Address, Hash, UInt } from '../utils/types';
 
 // interfaces need to be exported, and we need/want to support `import * as RaidenActions`
 type ChannelId = {
@@ -12,20 +11,15 @@ type ChannelId = {
 /* A new head in the blockchain is detected by provider */
 export const newBlock = createStandardAction('newBlock')<{ blockNumber: number }>();
 
-/* A new token network is detected in the TokenNetworkRegistry instance */
-export const tokenMonitored = createStandardAction('tokenMonitored').map(
-  ({
-    token,
-    tokenNetwork,
-    first = false,
-  }: {
-    token: Address;
-    tokenNetwork: Address;
-    first?: boolean;
-  }) => ({
-    payload: { token, tokenNetwork, first },
-  }),
-);
+/**
+ * A new token network is detected in the TokenNetworkRegistry instance
+ * fromBlock is only set on the first time, to fetch and handle past events
+ */
+export const tokenMonitored = createStandardAction('tokenMonitored')<{
+  token: Address;
+  tokenNetwork: Address;
+  fromBlock?: number;
+}>();
 
 /**
  * Channel actions receive ChannelId as 'meta' action property
@@ -34,13 +28,19 @@ export const tokenMonitored = createStandardAction('tokenMonitored').map(
 
 /* Request a channel to be opened with meta={ tokenNetwork, partner } and payload.settleTimeout */
 export const channelOpen = createStandardAction('channelOpen')<
-  { settleTimeout: number },
+  { settleTimeout?: number },
   ChannelId
 >();
 
 /* A channel is detected on-chain. Also works as 'success' for channelOpen action */
 export const channelOpened = createStandardAction('channelOpened')<
-  { id: number; settleTimeout: number; openBlock: number; txHash: Hash },
+  {
+    id: number;
+    settleTimeout: number;
+    openBlock: number;
+    isFirstParticipant: boolean;
+    txHash: Hash;
+  },
   ChannelId
 >();
 
@@ -57,13 +57,13 @@ export const channelMonitored = createStandardAction('channelMonitored')<
 
 /* Request a payload.deposit to be made to channel meta:ChannelId */
 export const channelDeposit = createStandardAction('channelDeposit')<
-  { deposit: BigNumber },
+  { deposit: UInt<32> },
   ChannelId
 >();
 
 /* A deposit is detected on-chain. Also works as 'success' for channelDeposit action */
 export const channelDeposited = createStandardAction('channelDeposited')<
-  { id: number; participant: Address; totalDeposit: BigNumber; txHash: Hash },
+  { id: number; participant: Address; totalDeposit: UInt<32>; txHash: Hash },
   ChannelId
 >();
 
@@ -71,6 +71,12 @@ export const channelDeposited = createStandardAction('channelDeposited')<
 export const channelDepositFailed = createStandardAction('channelDepositFailed').map(
   (payload: Error, meta: ChannelId) => ({ payload, error: true, meta }),
 );
+
+/* A withdraw is detected on-chain */
+export const channelWithdrawn = createStandardAction('channelWithdrawn')<
+  { id: number; participant: Address; totalWithdraw: UInt<32>; txHash: Hash },
+  ChannelId
+>();
 
 /* Request channel meta:ChannelId to be closed */
 export const channelClose = createStandardAction('channelClose')<undefined, ChannelId>();
