@@ -1,23 +1,24 @@
-import { Observable, from, of, EMPTY, merge, interval, ReplaySubject } from 'rxjs';
+import { Observable, from, of, EMPTY, merge, interval } from 'rxjs';
 import {
   catchError,
   filter,
   map,
   mergeMap,
   mergeMapTo,
-  startWith,
   tap,
   takeWhile,
   withLatestFrom,
   groupBy,
   exhaustMap,
-  multicast,
   first,
+  publishReplay,
+  switchMap,
+  pluck,
 } from 'rxjs/operators';
 import { isActionOf, ActionType } from 'typesafe-actions';
 import { findKey, get, isEmpty, negate } from 'lodash';
 
-import { BigNumber } from 'ethers/utils';
+import { BigNumber, hexlify, concat } from 'ethers/utils';
 import { Event } from 'ethers/contract';
 import { HashZero, Zero } from 'ethers/constants';
 
@@ -42,13 +43,15 @@ import {
   channelDeposited,
   channelClosed,
   channelSettled,
+  channelWithdrawn,
 } from './actions';
 import { SignatureZero, ShutdownReason } from '../constants';
-import { Address, Hash } from '../utils/types';
+import { Address, Hash, UInt, Signature } from '../utils/types';
 import { fromEthersEvent, getEventsStream, getNetwork } from '../utils/ethers';
+import { encode } from '../utils/data';
 
 /**
- * Register for new block events and emit newBlock actions for new blocks
+ * Fetch current blockNumber, register for new block events and emit newBlock actions
  *
  * @param action$ - Observable of RaidenActions
  * @param state$ - Observable of RaidenStates
@@ -57,12 +60,11 @@ import { fromEthersEvent, getEventsStream, getNetwork } from '../utils/ethers';
  */
 export const initNewBlockEpic = (
   {  }: Observable<RaidenAction>,
-  state$: Observable<RaidenState>,
+  {  }: Observable<RaidenState>,
   { provider }: RaidenEpicDeps,
 ): Observable<ActionType<typeof newBlock>> =>
-  state$.pipe(
-    first(),
-    mergeMap(() => fromEthersEvent<number>(provider, 'block')),
+  from(provider.getBlockNumber()).pipe(
+    mergeMap(blockNumber => merge(of(blockNumber), fromEthersEvent<number>(provider, 'block'))),
     map(blockNumber => newBlock({ blockNumber })),
   );
 
@@ -75,35 +77,42 @@ export const initNewBlockEpic = (
  * @returns Observable of tokenMonitored actions
  */
 export const initMonitorRegistryEpic = (
-  {  }: Observable<RaidenAction>,
+  action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
   { registryContract, contractsInfo }: RaidenEpicDeps,
 ): Observable<ActionType<typeof tokenMonitored>> =>
   state$.pipe(
-    first(),
-    mergeMap(state =>
-      merge(
-        // monitor old (in case of empty tokens) and new registered tokens
-        // and starts monitoring every registered token
-        getEventsStream<[Address, Address, Event]>(
-          registryContract,
-          [registryContract.filters.TokenNetworkCreated(null, null)],
-          isEmpty(state.tokens) ? of(contractsInfo.TokenNetworkRegistry.block_number) : undefined,
-          isEmpty(state.tokens) ? of(state.blockNumber) : undefined,
-        ).pipe(
-          withLatestFrom(state$.pipe(startWith(state))),
-          map(([[token, tokenNetwork], state]) =>
-            tokenMonitored({
-              token,
-              tokenNetwork,
-              first: !(token in state.tokens),
-            }),
-          ),
-        ),
-        // monitor previously monitored tokens
-        from(Object.entries(state.tokens)).pipe(
-          map(([token, tokenNetwork]) =>
-            tokenMonitored({ token: token as Address, tokenNetwork: tokenNetwork as Address }),
+    publishReplay(1, undefined, state$ =>
+      action$.pipe(
+        first(isActionOf(newBlock)),
+        withLatestFrom(state$),
+        switchMap(([{ payload: { blockNumber } }, state]) =>
+          merge(
+            // monitor old (in case of empty tokens) and new registered tokens
+            // and starts monitoring every registered token
+            getEventsStream<[Address, Address, Event]>(
+              registryContract,
+              [registryContract.filters.TokenNetworkCreated(null, null)],
+              isEmpty(state.tokens)
+                ? of(contractsInfo.TokenNetworkRegistry.block_number)
+                : undefined,
+              isEmpty(state.tokens) ? of(blockNumber) : undefined,
+            ).pipe(
+              withLatestFrom(state$),
+              map(([[token, tokenNetwork, event], state]) =>
+                tokenMonitored({
+                  token,
+                  tokenNetwork,
+                  fromBlock: !(token in state.tokens) ? event.blockNumber : undefined,
+                }),
+              ),
+            ),
+            // monitor previously monitored tokens
+            from(Object.entries(state.tokens)).pipe(
+              map(([token, tokenNetwork]) =>
+                tokenMonitored({ token: token as Address, tokenNetwork }),
+              ),
+            ),
           ),
         ),
       ),
@@ -146,12 +155,10 @@ export const initMonitorChannelsEpic = (
  */
 export const initMonitorProviderEpic = (
   {  }: Observable<RaidenAction>,
-  state$: Observable<RaidenState>,
+  {  }: Observable<RaidenState>,
   { address, network, provider }: RaidenEpicDeps,
 ): Observable<ActionType<typeof raidenShutdown>> =>
-  state$.pipe(
-    first(),
-    mergeMap(() => provider.listAccounts()),
+  from(provider.listAccounts()).pipe(
     // at init time, check if our address is in provider's accounts list
     // if not, it means Signer is a local Wallet or another non-provider-side account
     // if yes, poll accounts every 1s and monitors if address is still there
@@ -202,11 +209,11 @@ export const initMonitorProviderEpic = (
 export const tokenMonitoredEpic = (
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
-  { address, getTokenNetworkContract, contractsInfo }: RaidenEpicDeps,
+  { address, getTokenNetworkContract }: RaidenEpicDeps,
 ): Observable<ActionType<typeof channelOpened>> =>
   state$.pipe(
-    map(state => state.blockNumber),
-    multicast(new ReplaySubject(1), blockNumber$ =>
+    pluck('blockNumber'),
+    publishReplay(1, undefined, blockNumber$ =>
       action$.pipe(
         filter(isActionOf(tokenMonitored)),
         groupBy(action => action.payload.tokenNetwork),
@@ -224,16 +231,13 @@ export const tokenMonitoredEpic = (
                 tokenNetworkContract.filters.ChannelOpened(null, null, address, null),
               ];
 
-              console.log('getEventsStream', action, blockNumber$);
               return getEventsStream<ChannelOpenedEvent>(
                 tokenNetworkContract,
                 filters,
                 // if first time monitoring this token network,
                 // fetch TokenNetwork's pastEvents since registry deployment as fromBlock$
-                action.payload.first
-                  ? of(contractsInfo.TokenNetworkRegistry.block_number)
-                  : undefined,
-                action.payload.first ? blockNumber$ : undefined,
+                action.payload.fromBlock ? of(action.payload.fromBlock) : undefined,
+                action.payload.fromBlock ? blockNumber$ : undefined,
               ).pipe(
                 filter(([, p1, p2]) => p1 === address || p2 === address),
                 map(([id, p1, p2, settleTimeout, event]) =>
@@ -242,6 +246,7 @@ export const tokenMonitoredEpic = (
                       id: id.toNumber(),
                       settleTimeout: settleTimeout.toNumber(),
                       openBlock: event.blockNumber!,
+                      isFirstParticipant: address === p1,
                       txHash: event.transactionHash! as Hash,
                     },
                     {
@@ -278,11 +283,16 @@ export const channelMonitoredEpic = (
   state$: Observable<RaidenState>,
   { getTokenNetworkContract }: RaidenEpicDeps,
 ): Observable<
-  ActionType<typeof channelDeposited | typeof channelClosed | typeof channelSettled>
+  ActionType<
+    | typeof channelDeposited
+    | typeof channelWithdrawn
+    | typeof channelClosed
+    | typeof channelSettled
+  >
 > =>
   state$.pipe(
-    map(state => state.blockNumber),
-    multicast(new ReplaySubject(1), blockNumber$ =>
+    pluck('blockNumber'),
+    publishReplay(1, undefined, blockNumber$ =>
       action$.pipe(
         filter(isActionOf(channelMonitored)),
         groupBy(
@@ -295,11 +305,13 @@ export const channelMonitoredEpic = (
 
               // type of elements emitted by getEventsStream (past and new events coming from
               // contract): [channelId, participant, totalDeposit, Event]
-              type ChannelNewDepositEvent = [BigNumber, Address, BigNumber, Event];
-              // [channelId, participant, nonce, Event]
-              type ChannelClosedEvent = [BigNumber, Address, BigNumber, Event];
-              // [channelId, participant1amount, participant2amount, Event]
-              type ChannelSettledEvent = [BigNumber, BigNumber, BigNumber, Event];
+              type ChannelNewDepositEvent = [BigNumber, Address, UInt<32>, Event];
+              // [channelId, participant, totalWithdraw, Event]
+              type ChannelWithdrawEvent = [BigNumber, Address, UInt<32>, Event];
+              // [channelId, participant, nonce, balanceHash, Event]
+              type ChannelClosedEvent = [BigNumber, Address, UInt<8>, Hash, Event];
+              // [channelId, part1_amount, part1_locksroot, part2_amount, part2_locksroot Event]
+              type ChannelSettledEvent = [BigNumber, UInt<32>, Hash, UInt<32>, Hash, Event];
 
               // TODO: instead of one filter for each event, optimize to one filter per channel
               // it requires ethers to support OR filters for topics:
@@ -310,13 +322,21 @@ export const channelMonitoredEpic = (
                   null,
                   null,
                 ),
+                withdrawFilter = tokenNetworkContract.filters.ChannelWithdraw(
+                  action.payload.id,
+                  null,
+                  null,
+                ),
                 closedFilter = tokenNetworkContract.filters.ChannelClosed(
                   action.payload.id,
+                  null,
                   null,
                   null,
                 ),
                 settledFilter = tokenNetworkContract.filters.ChannelSettled(
                   action.payload.id,
+                  null,
+                  null,
                   null,
                   null,
                 );
@@ -342,13 +362,31 @@ export const channelMonitoredEpic = (
                     ),
                   ),
                 ),
+                getEventsStream<ChannelWithdrawEvent>(
+                  tokenNetworkContract,
+                  [withdrawFilter],
+                  action.payload.fromBlock ? of(action.payload.fromBlock) : undefined,
+                  action.payload.fromBlock ? blockNumber$ : undefined,
+                ).pipe(
+                  map(([id, participant, totalWithdraw, event]) =>
+                    channelWithdrawn(
+                      {
+                        id: id.toNumber(),
+                        participant,
+                        totalWithdraw,
+                        txHash: event.transactionHash! as Hash,
+                      },
+                      action.meta,
+                    ),
+                  ),
+                ),
                 getEventsStream<ChannelClosedEvent>(
                   tokenNetworkContract,
                   [closedFilter],
                   action.payload.fromBlock ? of(action.payload.fromBlock) : undefined,
                   action.payload.fromBlock ? blockNumber$ : undefined,
                 ).pipe(
-                  map(([id, participant, , event]) =>
+                  map(([id, participant, , , event]) =>
                     channelClosed(
                       {
                         id: id.toNumber(),
@@ -366,7 +404,7 @@ export const channelMonitoredEpic = (
                   action.payload.fromBlock ? of(action.payload.fromBlock) : undefined,
                   action.payload.fromBlock ? blockNumber$ : undefined,
                 ).pipe(
-                  map(([id, , , event]) =>
+                  map(([id, , , , , event]) =>
                     channelSettled(
                       {
                         id: id.toNumber(),
@@ -381,7 +419,10 @@ export const channelMonitoredEpic = (
                 // takeWhile tends to broad input to generic Action. We need to narrow it by hand
                 takeWhile<
                   ActionType<
-                    typeof channelDeposited | typeof channelClosed | typeof channelSettled
+                    | typeof channelDeposited
+                    | typeof channelWithdrawn
+                    | typeof channelClosed
+                    | typeof channelSettled
                   >
                 >(negate(isActionOf(channelSettled)), true),
               );
@@ -407,12 +448,12 @@ export const channelMonitoredEpic = (
 export const channelOpenEpic = (
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
-  { getTokenNetworkContract }: RaidenEpicDeps,
+  { getTokenNetworkContract, config$ }: RaidenEpicDeps,
 ): Observable<ActionType<typeof channelOpenFailed>> =>
   action$.pipe(
     filter(isActionOf(channelOpen)),
-    withLatestFrom(state$),
-    mergeMap(([action, state]) => {
+    withLatestFrom(state$, config$),
+    mergeMap(([action, state, config]) => {
       const tokenNetwork = getTokenNetworkContract(action.meta.tokenNetwork);
       const channelState = get(state.channels, [
         action.meta.tokenNetwork,
@@ -430,7 +471,7 @@ export const channelOpenEpic = (
         tokenNetwork.functions.openChannel(
           state.address,
           action.meta.partner,
-          action.payload.settleTimeout,
+          action.payload.settleTimeout || config.settleTimeout,
         ),
       ).pipe(
         mergeMap(async tx => ({ receipt: await tx.wait(), tx })),
@@ -589,7 +630,7 @@ export const channelDepositEpic = (
 export const channelCloseEpic = (
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
-  { getTokenNetworkContract }: RaidenEpicDeps,
+  { address, network, getTokenNetworkContract, signer }: RaidenEpicDeps,
 ): Observable<ActionType<typeof channelCloseFailed>> =>
   action$.pipe(
     filter(isActionOf(channelClose)),
@@ -612,17 +653,49 @@ export const channelCloseEpic = (
       }
       const channelId = channel.id;
 
-      // send closeChannel transaction
-      return from(
-        tokenNetworkContract.functions.closeChannel(
-          channelId,
-          action.meta.partner,
-          HashZero,
-          0,
-          HashZero,
-          SignatureZero,
+      const balanceHash = HashZero as Hash,
+        nonce = Zero as UInt<8>,
+        additionalHash = HashZero as Hash,
+        nonClosingSignature = hexlify(SignatureZero) as Signature;
+
+      // TODO: enable this after we're able to receive transfers
+      // if (channel.partner.balanceProof) {
+      //   balanceHash = createBalanceHash(
+      //     channel.partner.balanceProof.transferredAmount,
+      //     channel.partner.balanceProof.lockedAmount,
+      //     channel.partner.balanceProof.locksroot,
+      //   );
+      //   nonce = channel.partner.balanceProof.nonce;
+      //   additionalHash = channel.partner.balanceProof.messageHash;
+      //   nonClosingSignature = channel.partner.balanceProof.signature;
+      // }
+
+      const closingMessage = concat([
+        encode(action.meta.tokenNetwork, 20),
+        encode(network.chainId, 32),
+        encode(1, 32), // raiden_contracts.constants.MessageTypeId.BALANCE_PROOF
+        encode(channelId, 32),
+        encode(balanceHash, 32),
+        encode(nonce, 32),
+        encode(additionalHash, 32),
+        encode(nonClosingSignature, 65), // partner's signature for this balance proof
+      ]); // UInt8Array of 277 bytes
+
+      // sign counter balance proof (while we don't receive transfers yet, it's always zero),
+      // then send closeChannel transaction with our signature
+      return from(signer.signMessage(closingMessage) as Promise<Signature>).pipe(
+        mergeMap(closingSignature =>
+          tokenNetworkContract.functions.closeChannel(
+            channelId,
+            action.meta.partner,
+            address,
+            balanceHash,
+            nonce,
+            additionalHash,
+            nonClosingSignature,
+            closingSignature,
+          ),
         ),
-      ).pipe(
         tap(tx =>
           console.log(`sent closeChannel tx "${tx.hash}" to "${action.meta.tokenNetwork}"`),
         ),
@@ -683,18 +756,33 @@ export const channelSettleEpic = (
       }
       const channelId = channel.id;
 
+      const zeroBalanceProof = {
+        transferredAmount: Zero as UInt<32>,
+        lockedAmount: Zero as UInt<32>,
+        locksroot: HashZero as Hash,
+      };
+      let part1 = {
+          address: action.meta.partner,
+          ...(channel.partner.balanceProof || zeroBalanceProof),
+        },
+        part2 = {
+          address,
+          ...(channel.own.balanceProof || zeroBalanceProof),
+        };
+      if (channel.isFirstParticipant) [part1, part2] = [part2, part1];
+
       // send settleChannel transaction
       return from(
         tokenNetworkContract.functions.settleChannel(
           channelId,
-          address,
-          Zero,
-          Zero,
-          HashZero,
-          action.meta.partner,
-          Zero,
-          Zero,
-          HashZero,
+          part1.address,
+          part1.transferredAmount,
+          part1.lockedAmount,
+          part1.locksroot,
+          part2.address,
+          part2.transferredAmount,
+          part2.lockedAmount,
+          part2.locksroot,
         ),
       ).pipe(
         tap(tx =>
