@@ -11,6 +11,11 @@ import {
   SecretRequest,
   Unlock,
   Signed,
+  Metadata,
+  ToDevice,
+  WithdrawRequest,
+  WithdrawConfirmation,
+  WithdrawExpired,
 } from 'raiden-ts/messages/types';
 import {
   packMessage,
@@ -20,6 +25,7 @@ import {
   decodeJsonMessage,
   getBalanceProofFromEnvelopeMessage,
   createMessageHash,
+  createMetadataHash,
 } from 'raiden-ts/messages/utils';
 import { Address, Hash, Secret, UInt } from 'raiden-ts/utils/types';
 import { bigNumberify } from 'ethers/utils';
@@ -45,7 +51,6 @@ describe('sign/verify, pack & encode/decode ', () => {
       recipient: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
       locksroot: '0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b' as Hash,
       lock: {
-        type: 'Lock',
         amount: bigNumberify(10) as UInt<32>,
         expiration: One as UInt<32>,
         secrethash: '0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8' as Hash,
@@ -53,22 +58,36 @@ describe('sign/verify, pack & encode/decode ', () => {
       target: '0x811957b07304d335B271feeBF46754696694b09e' as Address,
       initiator: '0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address,
       fee: Zero as UInt<32>,
+      metadata: {
+        routes: [
+          {
+            route: [
+              '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
+              '0x811957b07304d335B271feeBF46754696694b09e' as Address,
+            ],
+          },
+        ],
+      },
     };
 
+    expect(createMessageHash(message)).toEqual(
+      '0x095a9cd18a990af080bab703e5004602b13ac8e2e4295421b73bd99c3c778967',
+    );
+
     expect(packMessage(message)).toEqual(
-      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e0000000000000000000000000000000000000000000000000000000000000001219f8ba12d6dd5c4076af98d9b608ab10351294d4433fde115fbd23243b48306',
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e0000000000000000000000000000000000000000000000000000000000000001095a9cd18a990af080bab703e5004602b13ac8e2e4295421b73bd99c3c778967',
     );
 
     const signed = await signMessage(signer, message);
     expect(Signed(LockedTransfer).is(signed)).toBe(true);
     expect(signed.signature).toBe(
-      '0xda8405b8e7d56425ed06e3bd4339c47a1c5e7326f92ad59fd859fa8b4888c5e4223faeaa889098c1dd8fb35cd3e3e716eb6ad628f577f8acf9fb2a895cae436e1b',
+      '0x81a29e70b8f36379a3c0939f1e23c66a75a820dddfa87319d5022431cdc5ad0471281e13437ae58f0d3feb80f0193ede01f2721c82610c8f0b782ec723e85be21c',
     );
     expect(getMessageSigner(signed)).toBe(address);
 
     const encoded = encodeJsonMessage(signed);
     expect(encoded).toBe(
-      '{"type":"LockedTransfer","chain_id":337,"message_identifier":123456,"payment_identifier":1,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","token":"0xc778417E063141139Fce010982780140Aa0cD5Ab","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"recipient":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","lock":{"type":"Lock","amount":10,"expiration":1,"secrethash":"0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8"},"target":"0x811957b07304d335B271feeBF46754696694b09e","initiator":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","fee":0,"signature":"0xda8405b8e7d56425ed06e3bd4339c47a1c5e7326f92ad59fd859fa8b4888c5e4223faeaa889098c1dd8fb35cd3e3e716eb6ad628f577f8acf9fb2a895cae436e1b"}',
+      '{"type":"LockedTransfer","chain_id":337,"message_identifier":123456,"payment_identifier":1,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","token":"0xc778417E063141139Fce010982780140Aa0cD5Ab","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"recipient":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","lock":{"amount":10,"expiration":1,"secrethash":"0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8"},"target":"0x811957b07304d335B271feeBF46754696694b09e","initiator":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","fee":0,"metadata":{"routes":[{"route":["0x2A915FDA69746F515b46C520eD511401d5CCD5e2","0x811957b07304d335B271feeBF46754696694b09e"]}]},"signature":"0x81a29e70b8f36379a3c0939f1e23c66a75a820dddfa87319d5022431cdc5ad0471281e13437ae58f0d3feb80f0193ede01f2721c82610c8f0b782ec723e85be21c"}',
     );
 
     const decoded = decodeJsonMessage(encoded);
@@ -90,7 +109,6 @@ describe('sign/verify, pack & encode/decode ', () => {
       recipient: '0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address,
       locksroot: HashZero as Hash,
       lock: {
-        type: 'Lock',
         amount: bigNumberify(10) as UInt<32>,
         expiration: One as UInt<32>,
         secrethash: '0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8' as Hash,
@@ -98,22 +116,33 @@ describe('sign/verify, pack & encode/decode ', () => {
       target: '0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address,
       initiator: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
       fee: Zero as UInt<32>,
+      metadata: {
+        routes: [
+          {
+            route: ['0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address],
+          },
+        ],
+      },
     };
 
+    expect(createMessageHash(message)).toEqual(
+      '0x50e01fec6308b0a39230f2adf47ea697b2e581472760171b35d0a6a9ea8633bb',
+    );
+
     expect(packMessage(message)).toEqual(
-      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053ad11d651b5158961173ce2ce735c1d2ca57e8d784b9e3ad3451a446a09653fac200000000000000000000000000000000000000000000000000000000000000014d66a16b37edcbcb9d5d3253013b8789042a9c5b2a19ac8f84335b48ee7f05ba',
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053ad11d651b5158961173ce2ce735c1d2ca57e8d784b9e3ad3451a446a09653fac2000000000000000000000000000000000000000000000000000000000000000150e01fec6308b0a39230f2adf47ea697b2e581472760171b35d0a6a9ea8633bb',
     );
 
     const signed = await signMessage(signer, message);
     expect(Signed(RefundTransfer).is(signed)).toBe(true);
     expect(signed.signature).toBe(
-      '0xde9fd2dc3357a25fb843339e2d242192af84597da5112a36b98853da65b68ed05557604903610714f065bec3064b0b7fdf2f6b029db2615741c3b95c6931f8901c',
+      '0x7ec958b5d0accea573474e21298bcb38c35b48f54a1fafcf0beba7ed48e2a9f21f28742471124e3f15fee445c46b130756dc9d4115563b1a420266caec65c6c71c',
     );
     expect(getMessageSigner(signed)).toBe(address);
 
     const encoded = encodeJsonMessage(signed);
     expect(encoded).toBe(
-      '{"type":"RefundTransfer","chain_id":337,"message_identifier":123457,"payment_identifier":1,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","token":"0xc778417E063141139Fce010982780140Aa0cD5Ab","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"recipient":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","locksroot":"0x0000000000000000000000000000000000000000000000000000000000000000","lock":{"type":"Lock","amount":10,"expiration":1,"secrethash":"0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8"},"target":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","initiator":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","fee":0,"signature":"0xde9fd2dc3357a25fb843339e2d242192af84597da5112a36b98853da65b68ed05557604903610714f065bec3064b0b7fdf2f6b029db2615741c3b95c6931f8901c"}',
+      '{"type":"RefundTransfer","chain_id":337,"message_identifier":123457,"payment_identifier":1,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","token":"0xc778417E063141139Fce010982780140Aa0cD5Ab","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"recipient":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","locksroot":"0x0000000000000000000000000000000000000000000000000000000000000000","lock":{"amount":10,"expiration":1,"secrethash":"0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8"},"target":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","initiator":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","fee":0,"metadata":{"routes":[{"route":["0x540B51eDc5900B8012091cc7c83caf2cb243aa86"]}]},"signature":"0x7ec958b5d0accea573474e21298bcb38c35b48f54a1fafcf0beba7ed48e2a9f21f28742471124e3f15fee445c46b130756dc9d4115563b1a420266caec65c6c71c"}',
     );
 
     const decoded = decodeJsonMessage(encoded);
@@ -135,20 +164,24 @@ describe('sign/verify, pack & encode/decode ', () => {
       locksroot: '0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b' as Hash,
     };
 
+    expect(createMessageHash(message)).toEqual(
+      '0x28603014ed14910483d354ef0160767629dbaadd55403ad271319f9d439531f1',
+    );
+
     expect(packMessage(message)).toEqual(
-      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e0000000000000000000000000000000000000000000000000000000000000001a0bf3aa37ee11d243bee523a4b0898ff3489fbf90609a4f41ef852a2cf0a31f5',
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e000000000000000000000000000000000000000000000000000000000000000128603014ed14910483d354ef0160767629dbaadd55403ad271319f9d439531f1',
     );
 
     const signed = await signMessage(signer, message);
     expect(Signed(Unlock).is(signed)).toBe(true);
     expect(signed.signature).toBe(
-      '0xbfc28e6ff99c4a3db920576b853a5484059a118c4f9a9105ec92ca0b68d873b0418266a6a32a8bb65652b4a38beb7440d7ee79263dc6e9e9652ab2443f5605951b',
+      '0xd153bbef8ca5462e62ba5dceda7929fc2ee7d866e983c033afe59e7f7958b8d80c734d5fba96028e8fb689300ca3c6a47764c0c0d6fbfffca9cf70729efb8e7e1c',
     );
     expect(getMessageSigner(signed)).toBe(address);
 
     const encoded = encodeJsonMessage(signed);
     expect(encoded).toBe(
-      '{"type":"Secret","chain_id":337,"message_identifier":123457,"payment_identifier":1,"secret":"0x3bc51dd335dda4f6aee24b3f88d88c5ee0b0d43aea4ed25a384531ce29fb062e","nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","signature":"0xbfc28e6ff99c4a3db920576b853a5484059a118c4f9a9105ec92ca0b68d873b0418266a6a32a8bb65652b4a38beb7440d7ee79263dc6e9e9652ab2443f5605951b"}',
+      '{"type":"Unlock","chain_id":337,"message_identifier":123457,"payment_identifier":1,"secret":"0x3bc51dd335dda4f6aee24b3f88d88c5ee0b0d43aea4ed25a384531ce29fb062e","nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","channel_identifier":1338,"transferred_amount":0,"locked_amount":10,"locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","signature":"0xd153bbef8ca5462e62ba5dceda7929fc2ee7d866e983c033afe59e7f7958b8d80c734d5fba96028e8fb689300ca3c6a47764c0c0d6fbfffca9cf70729efb8e7e1c"}',
     );
 
     const decoded = decodeJsonMessage(encoded);
@@ -170,20 +203,24 @@ describe('sign/verify, pack & encode/decode ', () => {
       locksroot: '0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b' as Hash,
     };
 
+    expect(createMessageHash(message)).toEqual(
+      '0x0d18bb3681423c08a7709849bb4ac045fed99551830031922ff4457a7b5eea1a',
+    );
+
     expect(packMessage(message)).toEqual(
-      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e00000000000000000000000000000000000000000000000000000000000000015edbeebd4f2f7c97a51f07a83d39bbc8e72a18dd12ba2141929609e4735dd791',
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000053a1d9479b298eb0a60edaf962f4cf092465456ad7a0265dfe28a0fe3a2a8ecef4e00000000000000000000000000000000000000000000000000000000000000010d18bb3681423c08a7709849bb4ac045fed99551830031922ff4457a7b5eea1a',
     );
 
     const signed = await signMessage(signer, message);
     expect(Signed(LockExpired).is(signed)).toBe(true);
     expect(signed.signature).toBe(
-      '0xecd53238b2aaf0885b8317f1b840fbc4f8cb22fb39f712284bda21e9842df5ca2f821a007b7196d0f8bce4e002717c9a86195532a82c74e8f403ee6b8f3e12641c',
+      '0xdd978b011c21327d9228f8ef179646df9029e14d91df38c41da9ded4819c5e304e5a9d14fe861b475816679597b66b0b08edea0e65e9674abf03dd3876ad9d801c',
     );
     expect(getMessageSigner(signed)).toBe(address);
 
     const encoded = encodeJsonMessage(signed);
     expect(encoded).toBe(
-      '{"type":"LockExpired","chain_id":337,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","message_identifier":123457,"channel_identifier":1338,"secrethash":"0xfdd5831261497a4de31cb31d29b3cafe1fd2dfcdadf3c4a72ed0af9bb106934d","transferred_amount":0,"locked_amount":10,"recipient":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","signature":"0xecd53238b2aaf0885b8317f1b840fbc4f8cb22fb39f712284bda21e9842df5ca2f821a007b7196d0f8bce4e002717c9a86195532a82c74e8f403ee6b8f3e12641c"}',
+      '{"type":"LockExpired","chain_id":337,"nonce":1,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","message_identifier":123457,"channel_identifier":1338,"secrethash":"0xfdd5831261497a4de31cb31d29b3cafe1fd2dfcdadf3c4a72ed0af9bb106934d","transferred_amount":0,"locked_amount":10,"recipient":"0x540B51eDc5900B8012091cc7c83caf2cb243aa86","locksroot":"0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b","signature":"0xdd978b011c21327d9228f8ef179646df9029e14d91df38c41da9ded4819c5e304e5a9d14fe861b475816679597b66b0b08edea0e65e9674abf03dd3876ad9d801c"}',
     );
 
     const decoded = decodeJsonMessage(encoded);
@@ -321,7 +358,6 @@ describe('sign/verify, pack & encode/decode ', () => {
       recipient: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
       locksroot: '0x607e890c54e5ba67cd483bedae3ba9da9bf2ef2fbf237b9fb39a723b2296077b' as Hash,
       lock: {
-        type: 'Lock',
         amount: bigNumberify(10) as UInt<32>,
         expiration: One as UInt<32>,
         secrethash: '0x59cad5948673622c1d64e2322488bf01619f7ff45789741b15a9f782ce9290a8' as Hash,
@@ -329,6 +365,17 @@ describe('sign/verify, pack & encode/decode ', () => {
       target: '0x811957b07304d335B271feeBF46754696694b09e' as Address,
       initiator: '0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address,
       fee: Zero as UInt<32>,
+      metadata: {
+        routes: [
+          {
+            route: [
+              '0x540B51eDc5900B8012091cc7c83caf2cb243aa86' as Address,
+              '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
+              '0x811957b07304d335B271feeBF46754696694b09e' as Address,
+            ],
+          },
+        ],
+      },
     };
 
     const signed = await signMessage(signer, message);
@@ -344,5 +391,161 @@ describe('sign/verify, pack & encode/decode ', () => {
       signature: expect.any(String),
       sender: address,
     });
+  });
+
+  test('create a metadata hash', () => {
+    const metadata: Metadata = {
+      routes: [
+        {
+          route: [
+            '0x77952Ce83Ca3cad9F7AdcFabeDA85Bd2F1f52008' as Address,
+            '0x94622cC2A5b64a58C25A129d48a2bEEC4b65b779' as Address,
+          ],
+        },
+      ],
+    };
+
+    expect(createMetadataHash(metadata)).toEqual(
+      '0x24b7955a3be270fd6c9513737759f42741653e9e39d901f7e2f255cc71dd4ae5',
+    );
+  });
+
+  test('ToDevice', async () => {
+    const message: ToDevice = {
+      type: MessageType.TO_DEVICE,
+      message_identifier: bigNumberify(123456) as UInt<8>,
+    };
+
+    expect(packMessage(message)).toEqual('0x0e000000000000000001e240');
+
+    const signed = await signMessage(signer, message);
+    expect(Signed(ToDevice).is(signed)).toBe(true);
+    expect(signed.signature).toBe(
+      '0x26674d7687baf09d21185664daf85cf6a9f9671d61f00e85d32621a721c3b4e251197dafe01d6930f1d86177e119dd0948f206d352204eb986b286b6b76a541d1b',
+    );
+    expect(getMessageSigner(signed)).toBe(address);
+
+    const encoded = encodeJsonMessage(signed);
+    expect(encoded).toBe(
+      '{"type":"ToDevice","message_identifier":123456,"signature":"0x26674d7687baf09d21185664daf85cf6a9f9671d61f00e85d32621a721c3b4e251197dafe01d6930f1d86177e119dd0948f206d352204eb986b286b6b76a541d1b"}',
+    );
+
+    const decoded = decodeJsonMessage(encoded);
+    expect(decoded).toEqual(signed);
+
+    // test sign already signed message return original object
+    const signed2 = await signMessage(signer, signed);
+    expect(signed2).toBe(signed);
+  });
+
+  test('WithdrawRequest', async () => {
+    const message: WithdrawRequest = {
+      type: MessageType.WITHDRAW_REQUEST,
+      chain_id: bigNumberify(337) as UInt<32>,
+      message_identifier: bigNumberify(123456) as UInt<8>,
+      token_network_address: '0xe82ae5475589b828D3644e1B56546F93cD27d1a4' as Address,
+      channel_identifier: bigNumberify(1338) as UInt<32>,
+      participant: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
+      total_withdraw: bigNumberify('10000000000000000000') as UInt<32>,
+      nonce: bigNumberify(135) as UInt<8>,
+      expiration: bigNumberify(182811) as UInt<32>,
+    };
+
+    expect(packMessage(message)).toEqual(
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000053a2a915fda69746f515b46c520ed511401d5ccd5e20000000000000000000000000000000000000000000000008ac7230489e80000000000000000000000000000000000000000000000000000000000000002ca1b',
+    );
+
+    const signed = await signMessage(signer, message);
+    expect(Signed(WithdrawRequest).is(signed)).toBe(true);
+    expect(signed.signature).toBe(
+      '0x5e0326b79f9ef19d6224317d54d17a55b4e1ebfc4d962388876d4575c421c4d238d50a892cd5e48d648c31c4f6ec5cb3947511a4dfe80c539875d859b1f31a0e1c',
+    );
+    expect(getMessageSigner(signed)).toBe(address);
+
+    const encoded = encodeJsonMessage(signed);
+    expect(encoded).toBe(
+      '{"type":"WithdrawRequest","chain_id":337,"message_identifier":123456,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","channel_identifier":1338,"participant":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","total_withdraw":10000000000000000000,"nonce":135,"expiration":182811,"signature":"0x5e0326b79f9ef19d6224317d54d17a55b4e1ebfc4d962388876d4575c421c4d238d50a892cd5e48d648c31c4f6ec5cb3947511a4dfe80c539875d859b1f31a0e1c"}',
+    );
+
+    const decoded = decodeJsonMessage(encoded);
+    expect(decoded).toEqual(signed);
+
+    // test sign already signed message return original object
+    const signed2 = await signMessage(signer, signed);
+    expect(signed2).toBe(signed);
+  });
+
+  test('WithdrawConfirmation', async () => {
+    const message: WithdrawConfirmation = {
+      type: MessageType.WITHDRAW_CONFIRMATION,
+      chain_id: bigNumberify(337) as UInt<32>,
+      message_identifier: bigNumberify(123456) as UInt<8>,
+      token_network_address: '0xe82ae5475589b828D3644e1B56546F93cD27d1a4' as Address,
+      channel_identifier: bigNumberify(1338) as UInt<32>,
+      participant: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
+      total_withdraw: bigNumberify('10000000000000000000') as UInt<32>,
+      nonce: bigNumberify(135) as UInt<8>,
+      expiration: bigNumberify(182811) as UInt<32>,
+    };
+
+    expect(packMessage(message)).toEqual(
+      '0xe82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000053a2a915fda69746f515b46c520ed511401d5ccd5e20000000000000000000000000000000000000000000000008ac7230489e80000000000000000000000000000000000000000000000000000000000000002ca1b',
+    );
+
+    const signed = await signMessage(signer, message);
+    expect(Signed(WithdrawConfirmation).is(signed)).toBe(true);
+    expect(signed.signature).toBe(
+      '0x5e0326b79f9ef19d6224317d54d17a55b4e1ebfc4d962388876d4575c421c4d238d50a892cd5e48d648c31c4f6ec5cb3947511a4dfe80c539875d859b1f31a0e1c',
+    );
+    expect(getMessageSigner(signed)).toBe(address);
+
+    const encoded = encodeJsonMessage(signed);
+    expect(encoded).toBe(
+      '{"type":"WithdrawConfirmation","chain_id":337,"message_identifier":123456,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","channel_identifier":1338,"participant":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","total_withdraw":10000000000000000000,"nonce":135,"expiration":182811,"signature":"0x5e0326b79f9ef19d6224317d54d17a55b4e1ebfc4d962388876d4575c421c4d238d50a892cd5e48d648c31c4f6ec5cb3947511a4dfe80c539875d859b1f31a0e1c"}',
+    );
+
+    const decoded = decodeJsonMessage(encoded);
+    expect(decoded).toEqual(signed);
+
+    // test sign already signed message return original object
+    const signed2 = await signMessage(signer, signed);
+    expect(signed2).toBe(signed);
+  });
+
+  test('WithdrawExpired', async () => {
+    const message: WithdrawExpired = {
+      type: MessageType.WITHDRAW_EXPIRED,
+      chain_id: bigNumberify(337) as UInt<32>,
+      message_identifier: bigNumberify(123456) as UInt<8>,
+      token_network_address: '0xe82ae5475589b828D3644e1B56546F93cD27d1a4' as Address,
+      channel_identifier: bigNumberify(1338) as UInt<32>,
+      participant: '0x2A915FDA69746F515b46C520eD511401d5CCD5e2' as Address,
+      total_withdraw: bigNumberify('10000000000000000000') as UInt<32>,
+      nonce: bigNumberify(135) as UInt<8>,
+      expiration: bigNumberify(182811) as UInt<32>,
+    };
+
+    expect(packMessage(message)).toEqual(
+      '0x110000000000000000000000000000000000000000000000000000000000000000000087000000000001e240e82ae5475589b828d3644e1b56546f93cd27d1a400000000000000000000000000000000000000000000000000000000000001510000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000053a2a915fda69746f515b46c520ed511401d5ccd5e20000000000000000000000000000000000000000000000008ac7230489e80000000000000000000000000000000000000000000000000000000000000002ca1b',
+    );
+
+    const signed = await signMessage(signer, message);
+    expect(Signed(WithdrawExpired).is(signed)).toBe(true);
+    expect(signed.signature).toBe(
+      '0xb4b2a079e00b9cfb05a3452500f052d3f1b549f3b3836c8698e6d6e71d65a34d2a52141562385b037b0a05f39484af153e84da84f6f507ebf83bc4f87183b7331c',
+    );
+    expect(getMessageSigner(signed)).toBe(address);
+
+    const encoded = encodeJsonMessage(signed);
+    expect(encoded).toBe(
+      '{"type":"WithdrawExpired","chain_id":337,"message_identifier":123456,"token_network_address":"0xe82ae5475589b828D3644e1B56546F93cD27d1a4","channel_identifier":1338,"participant":"0x2A915FDA69746F515b46C520eD511401d5CCD5e2","total_withdraw":10000000000000000000,"nonce":135,"expiration":182811,"signature":"0xb4b2a079e00b9cfb05a3452500f052d3f1b549f3b3836c8698e6d6e71d65a34d2a52141562385b037b0a05f39484af153e84da84f6f507ebf83bc4f87183b7331c"}',
+    );
+
+    const decoded = decodeJsonMessage(encoded);
+    expect(decoded).toEqual(signed);
+
+    // test sign already signed message return original object
+    const signed2 = await signMessage(signer, signed);
+    expect(signed2).toBe(signed);
   });
 });
